Moved Mail object to seperate package
